Sound Lures now exist
Created the Noise Component, attach it something that can make a noise. Call its GenerateNoise function. It has an array of feedback noises that get played randomly when the function is called and it also reports a noise event for the Kapre to be attracted to.

The BP_NoiseTrigger is just a box trigger that calls the GenerateNoise when it is stepped into.